# lib/api-zod/src/generated/types/createActivity.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Casa de Apoio - API de gestão
 * OpenAPI spec version: 0.1.0
 */
import type { CreateActivityStatus } from "./createActivityStatus";
import type { CreateActivityType } from "./createActivityType";

export interface CreateActivity {
  title: string;
  description?: string | null;
  type: CreateActivityType;
  status: CreateActivityStatus;
  scheduledDate: string;
  scheduledTime?: string | null;
  responsible?: string | null;
  residentId?: number | null;
  notes?: string | null;
}