# lib/api-zod/src/generated/types/resident.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Casa de Apoio - API de gestão
 * OpenAPI spec version: 0.1.0
 */
import type { ResidentStatus } from "./residentStatus";

export interface Resident {
  id: number;
  name: string;
  cpf: string;
  dateOfBirth: string;
  phone: string;
  emergencyContact: string;
  emergencyPhone: string;
  entryDate: string;
  exitDate?: string | null;
  status: ResidentStatus;
  room?: string | null;
  monthlyFee: number;
  notes?: string | null;
  createdAt: string;
}